Revert "Formally expose VSB_tofile in libvarnishapi"
This reverts commit 795366627ffdb0515c8e897a5f3d15623109cc56.

It shouldn't be exposed in libvarnishapi the way it is in this change
because it creates two diverging symbol versions (for systems with
linker support for symbol versioning) without changing the soname,
which of course we want to avoid in the first place in a stable branch.

If the requirement comes up we can revisit this topic and ensure both
binary stability and compatility (upgrade path).

Refs #3268