mgt/linux_jail: Adding linux jail support
Adding support for an experimental specific linux jailing.
The basic functionality is the same as UNIX jailing, using
the same parameters and same semantic for cgroups and UIDs.

The biggest initial difference is the use of SECCOMP to filter allowed
system calls. The filters are split in three groups:

	1) Manager filter (mgt_filter parameter);
	2) VCC compiler filter (vcc_filter parameter);
	3) Worker filter (wrk_filter parameter);

Each of these parameters is a plain text file containing the allowed
syscall numbers, one per line.

The default behavior of Linux jailing is killing the process which
is trying to execute a disallowed syscall with SIGSYS signal sent by
the kernel. In order to allow one to profile and create its own filter
there's an audit mode enabled by the parameter 'audit'.

The audit mode allow the execution of any syscall but it logs the call
to Linux default audit.log file. This can futher by queried by using
ausearch command.

This patch also includes standard filters for each process group as
example, they are located at /usr/share/doc/varnish. Those filters
are based on the default VCL shipped as example.

To start varnishd using the Linux jailing mechanism the following
command may be used:

  -j linux_experimental,mgt_filter=/usr/share/doc/varnish/wrk_filter,
  vcc_filter=/usr/share/doc/varnish/vcc_filter,
  wrk_filter=/usr/share/doc/varnish/wrk_filter

Signed-off-by: Marco Benatto <mbenatto@redhat.com>