Only build manual pages in maintainer mode
When we release a dist archive, it comes with the documentation but
since some of it is generated by programs we end up it makes no real
difference, and actually it makes things worse: rebuilding from the
dist archive ends up with a rebuild of the documentation because some
sources ended up being rebuilt. In other words, the documentation we
ship is dead weight in the dist archive and doesn't reduce the number
of build dependencies downstream.

From now on, rst2man remains mandatory to build our manual pages but
can safely be omitted by packaging scripts.