Revert "Only build manual pages in maintainer mode"
This reverts commit cc42bd31af078b8d0a2532d3744d2e662b16e422.

For some reason Travis CI jobs use autogen.sh and don't build with all
the bells and whistles we expect developers to work with. Despite the
explicit mention in autogen.des:

    # Use this when doing code development

Some of our continuous integration doesn't emulate this properly. At
least on the VTEST and CircleCI sides we do things as expected. I will
reintroduce this change later when Travis CI is ready to take it.