Revert "Revert "Only build manual pages in maintainer mode""
This brings back commit cc42bd31af078b8d0a2532d3744d2e662b16e422:

When we release a dist archive, it comes with the documentation but
since some of it is generated by programs it makes no real difference,
and actually it makes things worse: rebuilding from the dist archive
ends up with a rebuild of the documentation because some documentation
sources ended up being rebuilt. In other words, the documentation we
ship is dead weight in the dist archive and doesn't reduce the number
of build dependencies downstream.

From now on, rst2man remains mandatory to build our manual pages but
can safely be omitted by packaging scripts.