Revert "Revert "Revert "Only build manual pages in maintainer mode"""
This reverts commit 970c59edac07ccdb68c1ae13d94516ab8ea8fdfa.

this has been discussed on irc.

I agree that not rebuilding man pages when they are already there makes
sense.

But not building man pages when we could, does not make any sense to me.
And enforcing maintainer mode to build from git does not seem to make
any sense to me either.