build: Enable automake's subdir-objects option
We can get rid of autogen.des' egrep command used to silence this
warning too. In fact we can even remove the subshell and simply
call autoreconf directly. The problem with egrep is the loss of
a meaningful exit status for the autoreconf invocation.

I also enabled autoreconf's verbose output.