build: Facilitate VPATH bootstraps
In order to bootstrap a $(builddir) outside of $(srcdir) until now the
first step needed was to populate the environment. If for example one
wants to work from a build/ sub-directory the shortest autogen.des
invocation would look like:

    SRCDIR=.. ../autogen.des [configure options...]

Since we can infer $(srcdir) from autogen.des' relative path, it is now
possible to simply do this:

    ../autogen.des [configure options...]

It is still possible to provide a specific SRCDIR, even to a different
directory but the default should now work out of the box.