Complete the VSL memory fence pair
The log readers would not complete the acquire/release pair of memory
fence operations when waiting for new data to appear in the shared memory
log. With this, a VRMB() is issued after having waited for the
VSL_ENDMARKER disappears, and before using any data after it (including
the deciphered length data).