doc: It is now possible to name PRIV_ arguments
Contrary to what the documentation stated, PRIV_ arguments aren't
unnamed by nature with the current vmodtool script. Both the generated
struct and JSON descriptor would account for named PRIV_ arguments and
the only limitation was in libvcc.

It was really a bug in the sense that vmodtool would generate C code
incompatible with the code generated by libvcc. It was simpler to fix
libvcc and allow VMOD authors to give meaningful names to PRIVs if they
happen to use them in conjunction with [optional arguments].

Refs 82491c2a8abd