use a real target for the `$(srcdir)` to `$(builddir)` links for sphinx
The phony target was executed unconditionally and would thus race.

A real target (using `index.rst`) does not race and has the additional
advantage of simplifying the code (avoiding a condition in the shell
code).

It seems (at least GNU) `make` outsmarts the attempt to use a file from
`$(srcdir)` as a marker for the link to have succeeded when
`$(builddir)/index.rst` is used as the target.
Using `$(abs_builddir)` seems to dtrt.

Largely unrelated to this issue: The links are required for targets in
`BUILT_SOURCES`, so previously I added a dependency to the `all` target.
Having the link target in `BUILD_SOURCES` seems to be the better option.

Fixes #3370 (I hope)