more sphinx build fiddling
In the previous commit, I argued that adding the link job to
`BUILT_SOURCES` was the better option than adding it to the `all`
target, but I overlooked the fact that also the `BUILT_SOURCES` are
built in parallel (only before all other targets). Thus, we still need
an explicit dependency on the link job.