cli: New vcl.deps [-j] command
The plain text output is a suitable input for tsort(1), initially
meant to find the correct discard order, I realized that it would
be more reliable to let vcl.discard figure that one out.

It wouldn't really work in my initial scenario:

    varnishadm vcl.deps |
    <filter some VCLs out> |
    tsort |
    xargs -n 1 varnishadm vcl.discard

The filtering part only worked for direct dependencies but we can
have two levels with return(vcl).

I'm keeping this command only for setup analysis, the output can
easily be turned into a graphviz dot file for example.