vsc: New 'F' argument to force field inclusion
We need some fields to always be present in some cases with varnishstat
and as a result it disturbs end users settings. The new 'F' VSC argument
adds include patterns taking precedence over 'f' settings and is meant
for internal VUT usage.

Since varnishstat already needs to force some fields in curses mode, it's
the first consumer of the new argument.

Fixes #3394