vbe: Don't recycle Connection:close sessions
When a user adds a Connection:close header to the bereq, we end up
recycling the session and relying solely on the backend to close
it. This is not ideal when we deal with a misbehaving backend.

Conversely, if a backend replies with a Connection:close header,
or if a user manually adds it to the beresp, we should also not be
recycling sessions.