v1l: simplify workspace reservation handling
For niov == 0, we reserve up to IOV_MAX io vectors, otherwise niov io
vectors capped to IOV_MAX.

We used to handle the two cases differently and delay the WS_Release for
niov == 0 until V1L_Close(). There is no need to do so, in both cases we
can just reduce the reservation (turned into allocation) to the amount
of space we actually use.

As we are rolling back the workspace in V1L_Close(), this neither has
any influence on the state after the close.