cli: Teach vcl.discard to operate on multiple VCLs
To put it simply, let's take a simple CLI script:

    vcl.discard vcl1
    vcl.discard vcl2
    [...]
    vcl.discard vclX

We can now achieve the same with a single command:

    vcl.discard vcl1 vcl2 ... vclX

But there is slighty more to it, because vcl.discard operates on both
VCLs and VCL labels, and dependencies can exist between both. So in
addition to operate on multiple VCLs it also does so in the correct
order.