Fix a Flexelint warning which has bothered me for ages.
The message makes very little sense when you read it, and it was only
a pure accident which finally resolved what it means:

    Info 768: global struct member 'backend' (line 51, file cache/cache_backend.h) not referenced
    cache/cache_backend.h  51  Info 830: Location cited in prior message

If you have:

	struct backend;

	void somefunc(struct backend *);

	struct backend {
		...
	};

All is fine.

But without the `somefunc` prototype, the first vacuous mention of
`struct backend` is unused, and FlexeLint uses the name *of* the
struct as the name of a field *in* the struct, and compounds the
confusion by emitting the message on the line where struct backend
is finally flushed out.