sphinx: Accomodate VPATH builds
It is not possible to give sphinx-build multiple source directories, and
instead of trying to link missing files in $(builddir) to their $(srcdir)
counterpart we can merge them both directly in a new directory. But we
can't also merge in a way that would overwrite files:

    cp -Rf $(srcdir) $(builddir) <mergedir>

During a distcheck, on systems that support it, the source directory is
write-protected, and to some extent so are build directories. So instead
we need to proceed the other way around from $(builddir) to $(srcdir)
where files are copied in the merge directory only if there isn't already
one.

To spice things up a notch, we also need sources from a different tree,
and we shouldn't recursively copy $(builddir)/build in the merge
directory.

As it is not practical to encode all those rules into a single make
target it is moved to a dedicated script.

Refs #3374