param: Introduce the notion of PCRE parameters
Unlike string or VCC parameters they are inherited, but their definition
is not as straightforward as simple parameters. They also need to be
guarded by PARAM_ALL because inside struct params they are groupped in a
dedicated struct vre_limits, and there is no need for disruption in that
area.

Refs #3250