Introduce req.hash_always_pass
The current way to bypass the cache is to either `return(pass)` in
`vcl_recv` or to trigger an HfP/HfM on the backend side.

The former forces you to interrupt the request processing, such as
header normalization, or to set a variable/header to remember your
choice so that you can return `pass` once processing is done.

The latter is nicer but has to deal with request coalescing and it
doesn't really make sense to wait that long before make this decision.

`req.hash_always_pass` mirrors `req.hash_always_miss` but really just
transforms a `return(hash)` into a `return(pass)`. This means that it
has priority over `req.hash_always_miss` but also operate slightly
earlier, at the end of vcl_recv` rather than at the end of `vcl_hash`