vsl: Only collect candidate transactions
For high-throughput scenarios we might be collecting irrelevant
transactions, especially sessions that are never buffered, and
drastically increase the risk of overruns with long-running
sessions (typically when varnishd is behind a load balancer or
a TLS terminator that we can entrust with high client connection
reuse).

The -b and -c options have an effect on the output, which is way
too late for such a setup. These options primarily work with the
MSB backend and client markers but that is not enough to rule out
sessions since they are marked as client transactions. The same
goes for incomplete transactions for which we are lacking a Begin
tag, that would still be ruled out at the output stage.

We also need to collect irrelevant transactions for the sake of
grouping, because even when we do not wish to output them we have
to maintain the hierarchy.

Therefore, we can identify candidate transactions early and not
waste memory footprint and churn keeping track of transactions
that would have no effect besides increasing the risk of overruns.