Methods for vmod_priv
This is the refactoring we agreed on to enable an alternative
implementation of #3454. This PR does not yet introduce the copy
callback needed to add the functionality suggested in #3454.

We replace the .free pointer of struct vmod_priv with a pointer to a
methods struct with callbacks. As of this commit, it only contains
the .free callback renamed to .fini. The purpose of the refactoring is
to allow addition of more callbacks later.

The new struct vmod_priv_methods also contains a .type member pointing
to a string to contain an arbitrary description of the type of data any
priv holds which uses these methods.

Implementation:

relevant changes are in cache_vrt_priv.c and vrt.h, other changes are to
the documentation and bundled vmods.

The implementation is a simple refactoring for indirection of the call
to the .fini callback via the methods structure.