gc duplicate call to VBP_Update_Backend() when creating backends
VRT_new_backend_clustered() is the only caller of VBP_Insert().

There, VBP_Update_Backend() needs to be called after the director is
created in order to update the director status for a cold VCL.

For a warm VCL, VBP_Update_Backend() is called via
VRT_AddDirector() -> VDI_Event() -> vbe_dir_event() -> VBP_Control()

Thus, the additonal call before the director is added to the backend,
does not do anything but update the poll bits.

Somehow related to #3362