PoC: vcc: Allow arbitrary subs in the built-in VCL
Now that we know in advance the available suboutines in the built-in VCL
we can use that as the condition to create an append-able subroutine
when the vcl_ prefix is encountered.

This work is currently incomplete, at least because non-state built-in
subroutines are currently not listed when an unknwon subroutine is
encountered.

Also, instead of proper test coverage this currently uses b00000.vtc as
a strawman to show how to skip a specific processing.