vtc: Flip the feature check in v4
In order to run the test we want to NOT be using gcov, so we should test
that the environment variable is empty. In theory that should mean that
GCOV runs should currently fail since the feature check is currently
backwards so I may revert this change if that suddenly happens.

I find `test -z ...` more readable than `[ -z ... ]` in this context, so
if it comes to a revert, I will keep the test(1) syntax.