Avoid a second rbtree lookup in VRT_priv_task()
For a red-black tree insert in vrt_priv_dynamic(), we performed a lookup
and then an insert, resulting in basically the same tree traversal
twice.

We now prepare a new struct vrt_priv on the workspace and release it if
the respective priv already existed on the tree.

For a full workspace, we fall back to a lookup via
vrt_priv_dynamic_get().