vsl: Mask ExpKill logs by default
Based on anecdotal evidence that it is not often used, and the concern
of lock contention it might contribute to since such records are emitted
unbuffered for every single expiry event, it is now disabled by default.

ExpKill records are still available if needed for example for debugging
purposes, and are enabled in the 3 test cases where we use them. They
were also getting in the way of 1 test.

Closes #3475