varnishncsa: Skip irrelevant transactions (again)
This was already the behavior before #3468, but then the filtering was
moved to libvarnishapi to prevent varnishncsa from duplicating the logic
behind transaction collection.

Unfortunately, despite libvarnishapi's effort to not consider a client or
backend transaction without a Begin tag as a candidate, a transaction
missing one will get a synthetic Begin prior to dispatch.

Regardless, that would have only applied to vxid grouping so even if
varnishncsa was presented by default with only relevant transactions, it
could still see incomplete transactions in request grouping panic for
the same reason.

Fixes #3501