builtin: Split vcl_backend_response too
It's one more indirection than vcl_recv because the same logic applies
to multiple cases, so we first have subs for the different conditions
and they all call the shared vcl_beresp_hitmiss sub.

The vcl_beresp_hitmiss sub returns on purpose, that doesn't change the
default VCL behavior and makes it more usable from VCL code:

    if (some user condition applies) {
            call vcl_beresp_hitmiss;
    }

No need to force all call sites to return(deliver) when the desired
outcome is explicit. Extracting this sub also enables VCL authors to
only tweak the hitmiss TTL in one place.

For better compliance, we might also introduce a vcl_beresp_hitpass sub
for some of the built-in cases where a hit-for-miss object might be less
appropriate. The compliance effort will however change some semantics
and is outside of the scope of this split.