vcc: Repurpose (struct proc).called and add info to struct vcl_sub
The .called attribute was not used.

We now increment it only for custom subs which are actually called
statically (with "call"), that is, not for the builtin subs. This is
to simplify and clarify the use case in the next commits.

The increment happens for each VCC walk of the SUBs.

We also add the number of symbolic references and calls to (struct
vcl_sub) of the VGC to facilitate reviews and debugging.