Add disarmed assertions on ctx->method to vgc
in the course of #3163 we want to add a preamble to VCL subs, which,
ideally, would perform as many checks based on constants as possible.

This is a first step towards such checks: We add a commented-out
assert() on ctx->method for built-in subs only. To add the same for
custom subs, we will need the okmask from #3163, for which I would first
want to coordinate with phk.

Also, I would want to coordinate with him on arming the assertion, as
there might be a good reason why assert() is not yet defined for vgc (it
is already used in unused macros, though).

Having assertions would help humans, flexelint and compilers alike. Until
armed, the assertion helps humans at least.

The resulting vgc looks like this:

void v_matchproto_(vcl_func_f)
VGC_function_vcl_backend_error(VRT_CTX)
{
  // assert(ctx->method == (VCL_MET_BACKEND_ERROR));

  /* ... from ('Builtin' Line 165 Pos 23) */