vcc: Mark subs statically reachable via dynamic subs as "dynamic"
In follow up commits, we will need a marker for subs which can be
reachable via dynamically referenced SUBs.

We use "more references than calls" is this marker.  When walking
the call trees, we add one reference for all subs which are called from
a SUB having this marker, thus marking those, too.