dot: Rename lookup and lookup2 states
While lookup2 is indeed part of cnt_lookup(), let's call it hit since it
represents vcl_hit. As for lookup, let's call it _lookup since it is an
internal state not running in VCL space.

I'm not checking in the SVG file on purpose, this doesn't change the
effective graph.

Better diff with the --word-diff --word-diff-regex=. options.