vcc: Keep track of the effective syntax
The syntax is defined by the first `vcl x.y;` statement, subsequent vcl
versions become effective as long as they are compatible with the actual
syntax.

We ensure that we can't include 4.x from 5.x, but it's currently not
possible to have test coverage because other conditions are kicking in
prior to this check. We could change the order of checks or wait until
we have to major VCL versions.